fixed bugs with imports of datasets
