+++ dsa_helpers/__init__.py
@@ -3,7 +3,7 @@ from .imread import imread
 from .imwrite import imwrite
 
 # Version of the dsa-helpers package
-__version__ = "1.1.5.dev6"
+__version__ = "1.1.5.dev10"
 
 # To avoid slow downs, do not allow from dsa_helpers import * to import anything.
 __all__ = []
