+++ pyproject.toml
@@ -38,7 +38,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
 ]
 keywords = ["dsa", "girder", "girder_client"]
-version = "v1.1.5.dev6"
+version = "v1.1.5.dev10"
 
 [project.urls]
 Repository = "https://github.com/Gutman-Lab/dsa-python-utils"
